Added file name as edit function parameter
Use case: let's say you have i18n resources in JSON files in folders
like /en_US, /de_DE and so on. It is convenient to put all of them in
single resource file, which can be statically linked with the rest of
application code.  To do that we you need to parse file name and put the
JSON contents in separate subtrees with locale name as a key. With this
change it's possible to pass file name to edit function for that
purpose.