Added customizer function for lodash.mergeWith (#21)
* Added ability to put lodash.mergeWith customizer function

lodash.mergeWith supports a customizer function. Currently there are only 2 ways this function can be modified in gulp-merge-json. This change allows anyone to put their own customizer function in there instead of using either contact or merge arrays.

* Re-added deprecate dependency

Accidentally removed this in testing and added it back again.

* Corrected Node.js 4 incompatibility in object deconstruction

Corrected object deconstruction in function argument incompatibility as this was breaking the build.

* Changed to ES5-style object deconstruction

Explicitly defined each value ES5-style where ES2015 object deconstruction was used as it is not compatible with Node.js 4.